[database/postgres] PostgreSQL database driver

+++ src/upgrades/1.4.6/delete_sessions.js
@@ -45,10 +45,12 @@ module.exports = {
 					], function (err) {
 						next(err);
 					});
-				} else {
+				} else if (db.client && db.client.collection) {
 					db.client.collection('sessions').deleteMany({}, {}, function (err) {
 						next(err);
 					});
+				} else {
+					next();
 				}
 			},
 		], callback);
